Remove possibility of error when deleting all images
Likely caused by sending a request to S3 to delete an empty list of files

+++ Pages/Images/Index.cshtml.cs
@@ -70,16 +70,19 @@ namespace AwsAspCore.Pages.Images
             }
             while (!string.IsNullOrEmpty(response.ContinuationToken));
 
-            await s3.DeleteObjectsAsync(new DeleteObjectsRequest
+            if (result.Count > 0)
             {
-                BucketName = bucketName,
-                Quiet = true,
-                Objects = result.Select(obj => new KeyVersion
+                await s3.DeleteObjectsAsync(new DeleteObjectsRequest
                 {
-                    Key = obj.Key,
-                })
-                .ToList(),
-            });
+                    BucketName = bucketName,
+                    Quiet = true,
+                    Objects = result.Select(obj => new KeyVersion
+                    {
+                        Key = obj.Key,
+                    })
+                    .ToList(),
+                });
+            }
 
             return RedirectToPage("./Index");
         }
